fix TODOs. fix reprocessFullBatch. fix egp tx size
