+++ pool/pool.go
@@ -245,8 +245,8 @@ func (p *Pool) StoreTx(ctx context.Context, tx types.Transaction, ip string, isW
 
 // ValidateBreakEvenGasPrice validates the effective gas price
 func (p *Pool) ValidateBreakEvenGasPrice(ctx context.Context, tx types.Transaction, preExecutionGasUsed uint64, gasPrices GasPrices) error {
-	// Get the tx gas price we will use in the egp calculation. If egp is disabled we will use a "simulated" tx gas price
-	txGasPrice, _ := p.effectiveGasPrice.GetTxAndL2GasPrice(tx.GasPrice(), gasPrices.L1GasPrice, gasPrices.L2GasPrice)
+	// Get the tx gas price we will use in the egp calculation. If egp is disabled we will use a "simulated" tx gas price and l2 gas price
+	txGasPrice, l2GasPrice := p.effectiveGasPrice.GetTxAndL2GasPrice(tx.GasPrice(), gasPrices.L1GasPrice, gasPrices.L2GasPrice)
 
 	breakEvenGasPrice, err := p.effectiveGasPrice.CalculateBreakEvenGasPrice(tx.Data(), txGasPrice, preExecutionGasUsed, gasPrices.L1GasPrice)
 	if err != nil {
@@ -267,9 +267,9 @@ func (p *Pool) ValidateBreakEvenGasPrice(ctx context.Context, tx types.Transacti
 	tmpFactor.Int(breakEvenGasPriceWithFactor)
 
 	if breakEvenGasPriceWithFactor.Cmp(txGasPrice) == 1 { // breakEvenGasPriceWithMargin > txGasPrice
-		// check against L2GasPrice now
-		L2GasPrice := big.NewInt(0).SetUint64(gasPrices.L2GasPrice)
-		if txGasPrice.Cmp(L2GasPrice) == -1 { // txGasPrice < gasPrices.L2GasPrice
+		// check against l2GasPrice now
+		biL2GasPrice := big.NewInt(0).SetUint64(l2GasPrice)
+		if txGasPrice.Cmp(biL2GasPrice) == -1 { // txGasPrice < l2GasPrice
 			// reject tx
 			reject = true
 		} else {
@@ -279,7 +279,7 @@ func (p *Pool) ValidateBreakEvenGasPrice(ctx context.Context, tx types.Transacti
 	}
 
 	log.Infof("egp-log: txGasPrice(): %v, breakEvenGasPrice: %v, breakEvenGasPriceWithFactor: %v, gasUsed: %v, reject: %t, loss: %v, L1GasPrice: %d, L2GasPrice: %d, Enabled: %t, tx: %s",
-		txGasPrice, breakEvenGasPrice, breakEvenGasPriceWithFactor, preExecutionGasUsed, reject, loss, gasPrices.L1GasPrice, gasPrices.L2GasPrice, p.cfg.EffectiveGasPrice.Enabled, tx.Hash().String())
+		txGasPrice, breakEvenGasPrice, breakEvenGasPriceWithFactor, preExecutionGasUsed, reject, loss, gasPrices.L1GasPrice, l2GasPrice, p.cfg.EffectiveGasPrice.Enabled, tx.Hash().String())
 
 	// Reject transaction if EffectiveGasPrice is enabled
 	if p.cfg.EffectiveGasPrice.Enabled && reject {
