+++ sequencer/batch.go
@@ -2,13 +2,15 @@ package sequencer
 
 import (
 	"context"
-	"errors"
+	"encoding/json"
 	"fmt"
 	"time"
 
+	"github.com/0xPolygonHermez/zkevm-node/event"
 	"github.com/0xPolygonHermez/zkevm-node/log"
 	"github.com/0xPolygonHermez/zkevm-node/sequencer/metrics"
 	"github.com/0xPolygonHermez/zkevm-node/state"
+	stateMetrics "github.com/0xPolygonHermez/zkevm-node/state/metrics"
 	"github.com/ethereum/go-ethereum/common"
 )
 
@@ -90,8 +92,7 @@ func (f *finalizer) setWIPBatch(ctx context.Context, wipStateBatch *state.Batch)
 		return nil, err
 	}
 
-	//TODO: Check wipStateBatch.BatchNumber > 0
-	//TODO: We retrieve prevStateBatch only to init the initialStateRoot of the wip batch. To avoid this We can update the initialStateRoot in the state.batch table for the wip batch
+	// Retrieve prevStateBatch to init the initialStateRoot of the wip batch
 	prevStateBatch, err := f.state.GetBatchByNumber(ctx, wipStateBatch.BatchNumber-1, dbTx)
 	if err != nil {
 		return nil, err
@@ -213,9 +214,8 @@ func (f *finalizer) finalizeBatch(ctx context.Context) {
 
 	var err error
 	f.wipBatch, err = f.closeAndOpenNewWIPBatch(ctx)
-	for err != nil { //TODO: we need to review is this for loop is needed or if it's better to halt if we have an error
-		log.Errorf("failed to create new WIP batch. Error: %s", err)
-		f.wipBatch, err = f.closeAndOpenNewWIPBatch(ctx)
+	if err != nil {
+		f.halt(ctx, fmt.Errorf("failed to create new WIP batch. Error: %s", err))
 	}
 
 	log.Infof("new WIP batch %d", f.wipBatch.batchNumber)
@@ -228,7 +228,6 @@ func (f *finalizer) closeAndOpenNewWIPBatch(ctx context.Context) (*Batch, error)
 		f.finalizeL2Block(ctx)
 	}
 
-	//TODO: we can use only 1 WG for L2 blocks pending to process/store instead to use 2
 	// Wait until all L2 blocks are processed
 	startWait := time.Now()
 	f.pendingL2BlocksToProcessWG.Wait()
@@ -242,9 +241,6 @@ func (f *finalizer) closeAndOpenNewWIPBatch(ctx context.Context) (*Batch, error)
 	log.Debugf("waiting for pending L2 blocks to be stored took: %s", endWait.Sub(startWait).String())
 
 	var err error
-	if f.wipBatch.finalStateRoot == state.ZeroHash { //TODO: this check makes sense?
-		return nil, errors.New("state root must have value to close batch")
-	}
 
 	// We need to process the batch to update the state root before closing the batch
 	if f.wipBatch.initialStateRoot == f.wipBatch.finalStateRoot {
@@ -432,4 +428,224 @@ func (f *finalizer) closeWIPBatch(ctx context.Context) error {
 	}
 
 	return nil
+}
+
+// maxTxsPerBatchReached checks if the batch has reached the maximum number of txs per batch
+func (f *finalizer) maxTxsPerBatchReached() bool {
+	if f.wipBatch.countOfTxs >= int(f.batchConstraints.MaxTxsPerBatch) {
+		log.Infof("closing batch: %d, because it reached the maximum number of txs.", f.wipBatch.batchNumber)
+		f.wipBatch.closingReason = state.BatchFullClosingReason
+		return true
+	}
+	return false
+}
+
+// reprocessFullBatch reprocesses a batch used as sanity check
+func (f *finalizer) reprocessFullBatch(ctx context.Context, batchNum uint64, initialStateRoot common.Hash, initialAccInputHash common.Hash, expectedNewStateRoot common.Hash) (*state.ProcessBatchResponse, error) {
+	reprocessError := func(batch *state.Batch) {
+		f.halt(ctx, fmt.Errorf("error reprocessing full batch (sanity check). Check previous errors in logs to know which was the cause"))
+
+		if batch == nil {
+			return
+		}
+
+		rawL2Blocks, err := state.DecodeBatchV2(batch.BatchL2Data)
+		if err != nil {
+			log.Errorf("[reprocessFullBatch] error decoding BatchL2Data for batch %d. Error: %s", batch.BatchNumber, err)
+			return
+		}
+
+		// Log batch detailed info
+		log.Infof("[reprocessFullBatch] BatchNumber: %d, InitialStateRoot: %s, ExpectedNewStateRoot: %s, GER: %s", batch.BatchNumber, initialStateRoot, expectedNewStateRoot, batch.GlobalExitRoot)
+		for i, rawL2block := range rawL2Blocks.Blocks {
+			for j, rawTx := range rawL2block.Transactions {
+				log.Infof("[reprocessFullBatch] BatchNumber: %d, block position: % d, tx position %d, tx hash: %s", batch.BatchNumber, i, j, rawTx.Tx.Hash())
+			}
+		}
+	}
+
+	log.Debugf("[reprocessFullBatch] reprocessing batch: %d, InitialStateRoot: %s, ExpectedNewStateRoot: %s, GER: %s", batchNum, initialStateRoot, expectedNewStateRoot)
+
+	batch, err := f.state.GetBatchByNumber(ctx, batchNum, nil)
+	if err != nil {
+		log.Errorf("[reprocessFullBatch] failed to get batch %d, err: %s", batchNum, err)
+		reprocessError(nil)
+		return nil, ErrGetBatchByNumber
+	}
+
+	caller := stateMetrics.DiscardCallerLabel
+	if f.cfg.SequentialReprocessFullBatch {
+		caller = stateMetrics.SequencerCallerLabel
+	}
+
+	// TODO: review this request for reprocess full batch
+	executorBatchRequest := state.ProcessRequest{
+		BatchNumber:       batch.BatchNumber,
+		L1InfoRoot_V2:     mockL1InfoRoot,
+		OldStateRoot:      initialStateRoot,
+		OldAccInputHash:   initialAccInputHash,
+		Transactions:      batch.BatchL2Data,
+		Coinbase:          batch.Coinbase,
+		TimestampLimit_V2: uint64(time.Now().Unix()),
+		ForkID:            f.state.GetForkIDByBatchNumber(batch.BatchNumber),
+		Caller:            caller,
+	}
+
+	var result *state.ProcessBatchResponse
+
+	result, err = f.state.ProcessBatchV2(ctx, executorBatchRequest, false)
+	if err != nil {
+		log.Errorf("[reprocessFullBatch] failed to process batch %d. Error: %s", batch.BatchNumber, err)
+		reprocessError(batch)
+		return nil, ErrProcessBatch
+	}
+
+	if result.ExecutorError != nil {
+		log.Errorf("[reprocessFullBatch] executor error when reprocessing batch %d, error: %s", batch.BatchNumber, result.ExecutorError)
+		reprocessError(batch)
+		return nil, ErrExecutorError
+	}
+
+	if result.IsRomOOCError {
+		log.Errorf("[reprocessFullBatch] failed to process batch %d because OutOfCounters", batch.BatchNumber)
+		reprocessError(batch)
+
+		payload, err := json.Marshal(executorBatchRequest)
+		if err != nil {
+			log.Errorf("[reprocessFullBatch] error marshaling payload: %s", err)
+		} else {
+			event := &event.Event{
+				ReceivedAt:  time.Now(),
+				Source:      event.Source_Node,
+				Component:   event.Component_Sequencer,
+				Level:       event.Level_Critical,
+				EventID:     event.EventID_ReprocessFullBatchOOC,
+				Description: string(payload),
+				Json:        executorBatchRequest,
+			}
+			err = f.eventLog.LogEvent(ctx, event)
+			if err != nil {
+				log.Errorf("[reprocessFullBatch] error storing payload: %s", err)
+			}
+		}
+
+		return nil, ErrProcessBatchOOC
+	}
+
+	if result.NewStateRoot != expectedNewStateRoot {
+		log.Errorf("[reprocessFullBatch] new state root mismatch for batch %d, expected: %s, got: %s", batch.BatchNumber, expectedNewStateRoot.String(), result.NewStateRoot.String())
+		reprocessError(batch)
+		return nil, ErrStateRootNoMatch
+	}
+
+	log.Infof("[reprocessFullBatch]: reprocess successfully done for batch %d", batch.BatchNumber)
+	return result, nil
+}
+
+// checkRemainingResources checks if the transaction uses less resources than the remaining ones in the batch.
+func (f *finalizer) checkRemainingResources(result *state.ProcessBatchResponse, tx *TxTracker) error {
+	usedResources := state.BatchResources{
+		ZKCounters: result.UsedZkCounters,
+		Bytes:      uint64(len(tx.RawTx)),
+	}
+
+	err := f.wipBatch.remainingResources.Sub(usedResources)
+	if err != nil {
+		log.Infof("current transaction exceeds the remaining batch resources, updating metadata for tx in worker and continuing")
+		start := time.Now()
+		f.worker.UpdateTxZKCounters(result.BlockResponses[0].TransactionResponses[0].TxHash, tx.From, usedResources.ZKCounters)
+		metrics.WorkerProcessingTime(time.Since(start))
+		return err
+	}
+
+	return nil
+}
+
+// isBatchResourcesExhausted checks if one of resources of the wip batch has reached the max value
+func (f *finalizer) isBatchResourcesExhausted() bool {
+	resources := f.wipBatch.remainingResources
+	zkCounters := resources.ZKCounters
+	result := false
+	resourceDesc := ""
+	if resources.Bytes <= f.getConstraintThresholdUint64(f.batchConstraints.MaxBatchBytesSize) {
+		resourceDesc = "MaxBatchBytesSize"
+		result = true
+	} else if zkCounters.UsedSteps <= f.getConstraintThresholdUint32(f.batchConstraints.MaxSteps) {
+		resourceDesc = "MaxSteps"
+		result = true
+	} else if zkCounters.UsedPoseidonPaddings <= f.getConstraintThresholdUint32(f.batchConstraints.MaxPoseidonPaddings) {
+		resourceDesc = "MaxPoseidonPaddings"
+		result = true
+	} else if zkCounters.UsedBinaries <= f.getConstraintThresholdUint32(f.batchConstraints.MaxBinaries) {
+		resourceDesc = "MaxBinaries"
+		result = true
+	} else if zkCounters.UsedKeccakHashes <= f.getConstraintThresholdUint32(f.batchConstraints.MaxKeccakHashes) {
+		resourceDesc = "MaxKeccakHashes"
+		result = true
+	} else if zkCounters.UsedArithmetics <= f.getConstraintThresholdUint32(f.batchConstraints.MaxArithmetics) {
+		resourceDesc = "MaxArithmetics"
+		result = true
+	} else if zkCounters.UsedMemAligns <= f.getConstraintThresholdUint32(f.batchConstraints.MaxMemAligns) {
+		resourceDesc = "MaxMemAligns"
+		result = true
+	} else if zkCounters.GasUsed <= f.getConstraintThresholdUint64(f.batchConstraints.MaxCumulativeGasUsed) {
+		resourceDesc = "MaxCumulativeGasUsed"
+		result = true
+	} else if zkCounters.UsedSha256Hashes_V2 <= f.getConstraintThresholdUint32(f.batchConstraints.MaxSHA256Hashes) {
+		resourceDesc = "MaxSHA256Hashes"
+		result = true
+	}
+
+	if result {
+		log.Infof("closing batch %d, because it reached %s limit", f.wipBatch.batchNumber, resourceDesc)
+		f.wipBatch.closingReason = state.BatchAlmostFullClosingReason
+	}
+
+	return result
+}
+
+// getConstraintThresholdUint64 returns the threshold for the given input
+func (f *finalizer) getConstraintThresholdUint64(input uint64) uint64 {
+	return input * uint64(f.cfg.ResourcePercentageToCloseBatch) / 100 //nolint:gomnd
+}
+
+// getConstraintThresholdUint32 returns the threshold for the given input
+func (f *finalizer) getConstraintThresholdUint32(input uint32) uint32 {
+	return uint32(input*f.cfg.ResourcePercentageToCloseBatch) / 100 //nolint:gomnd
+}
+
+// getUsedBatchResources returns the max resources that can be used in a batch
+func getUsedBatchResources(constraints state.BatchConstraintsCfg, remainingResources state.BatchResources) state.BatchResources {
+	return state.BatchResources{
+		ZKCounters: state.ZKCounters{
+			GasUsed:              constraints.MaxCumulativeGasUsed - remainingResources.ZKCounters.GasUsed,
+			UsedKeccakHashes:     constraints.MaxKeccakHashes - remainingResources.ZKCounters.UsedKeccakHashes,
+			UsedPoseidonHashes:   constraints.MaxPoseidonHashes - remainingResources.ZKCounters.UsedPoseidonHashes,
+			UsedPoseidonPaddings: constraints.MaxPoseidonPaddings - remainingResources.ZKCounters.UsedPoseidonPaddings,
+			UsedMemAligns:        constraints.MaxMemAligns - remainingResources.ZKCounters.UsedMemAligns,
+			UsedArithmetics:      constraints.MaxArithmetics - remainingResources.ZKCounters.UsedArithmetics,
+			UsedBinaries:         constraints.MaxBinaries - remainingResources.ZKCounters.UsedBinaries,
+			UsedSteps:            constraints.MaxSteps - remainingResources.ZKCounters.UsedSteps,
+			UsedSha256Hashes_V2:  constraints.MaxSHA256Hashes - remainingResources.ZKCounters.UsedSha256Hashes_V2,
+		},
+		Bytes: constraints.MaxBatchBytesSize - remainingResources.Bytes,
+	}
+}
+
+// getMaxRemainingResources returns the max zkcounters that can be used in a batch
+func getMaxRemainingResources(constraints state.BatchConstraintsCfg) state.BatchResources {
+	return state.BatchResources{
+		ZKCounters: state.ZKCounters{
+			GasUsed:              constraints.MaxCumulativeGasUsed,
+			UsedKeccakHashes:     constraints.MaxKeccakHashes,
+			UsedPoseidonHashes:   constraints.MaxPoseidonHashes,
+			UsedPoseidonPaddings: constraints.MaxPoseidonPaddings,
+			UsedMemAligns:        constraints.MaxMemAligns,
+			UsedArithmetics:      constraints.MaxArithmetics,
+			UsedBinaries:         constraints.MaxBinaries,
+			UsedSteps:            constraints.MaxSteps,
+			UsedSha256Hashes_V2:  constraints.MaxSHA256Hashes,
+		},
+		Bytes: constraints.MaxBatchBytesSize,
+	}
 }
